+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+### Features
+- add configuratable options
+
 
 <a name="v0.2.3"></a>
 ## [v0.2.3] - 2021-08-17